Revert "Memindahkan atribut dan fungsi salary menjadi satu class"
This reverts commit a67c2d1e6e50ca032263d6c269a2d425f629fa90.
